requirements.txt
streamlit
google-generativeai
requests
python-dotenv